### 2018-08-29 #920 @whedon
[PRE REVIEW]: bem: Modeling for Neutron Bragg-Edge Imaging
**Submitting author:** @Yxqd (<a href="http://orcid.org/0000-0001-9233-0100">Jiao Lin</a>)
**Repository:** <a href="https://github.com/ornlneutronimaging/braggedgemodeling" target ="_blank">https://github.com/ornlneutronimaging/braggedgemodeling</a>
**Version:** v0.1.1
