### 2018-08-30 #920 @katyhuff
@jwuttke @mdoucet @stuartcampbell @mikapfl @marcocamma : All of you seem to be eminently qualified to review this JOSS submission ["bem: Modeling for Neutron Bragg-Edge Imaging"](https://github.com/openjournals/joss-papers/blob/joss.00920/joss.00920/10.21105.joss.00920.pdf).

Would any of you be interested in reviewing it and available to do so soon?
