### 2018-08-30 #920 @katyhuff
Thanks @stuartcampbell for your prompt response! Hopefully one or two of the others will also be available in the next few weeks to look this over. ( @jwuttke @mdoucet @mikapfl @marcocamma )
