### 2018-08-30 #920 @katyhuff
@stuartcampbell Sorry I missed that -- Consider where you think you land in the COI policy. I think this might be a borderline case, but the assessment of COI is up to you https://github.com/openjournals/joss/blob/master/COI.md
