### 2018-08-30 #920 @yxqd
Thanks, guys. I have not yet collaborated with either @stuartcampbell or @mdoucet on any published works. But yes @mdoucet and I are in the same group, and @stuartcampbell and I overlapped for a bit more than a year at ORNL.
