### 2018-09-05 #920 @stuartcampbell
@katyhuff I think I am ok, I recently reviewed something with a similar COI and @arfon agreed I wasn't conflicted!
