### 2018-09-13 #920 @mikapfl
Hi,

I could do a review of the documentation and code, but I can't really say anything about the physics since I haven't worked with Neutrons yet.

Cheers,

Mika
