### 2018-09-17 ornlneutronimaging/braggedgemodeling#23 @mikapfl
Does not install via setup.py due to diffpy.Structure
Hi,

I tried installing via
```bash
python3 setup.py install --user
```
as suggested in the [docs](https://ornlneutronimaging.github.io/braggedgemodeling/installation.html#using-setup-py).
However, that failed because the `diffpy.Structure` package available from https://pypi.org/project/diffpy.Structure/ is not compatible with python3. Without it, `bem` is not importable. I think you have to add documentation about where I would acquire a `diffpy.Structure` version compatible to `bem` and how to install that.

Cheers,

Mika
