### 2018-09-25 #973 @yxqd
Thanks @katyhuff. Song made significant contribution to the software by providing the first implementation of the algorithm in Excel :) ; doing the research using the code; and provided numerous feedback.

I will fix the installation instructions.
