### 2018-09-25 46a5f27
Added all the preparation code. Code that has to be run before running matlab.
