+++ bem/texture/preparation/vdrive_handler.py
@@ -1,0 +1,452 @@
+import numpy as np
+import os
+import pandas as pd
+import re
+
+
+class Data(object):
+
+    # pandas objects
+    raw = None # just after loading the file
+    raw_iv = None # just I/V columns data
+    raw_eiv = None # just eI/V columns data
+    cleaned = None  # only columns of interest
+    bank1 = None # only data from bank1
+    bank2 = None # only data from bank2
+
+    filename = ''
+
+class Bank(object):
+
+    omega = []
+    hrot = []
+    psi = []
+    phi = []
+
+    iv = []  # pandas object
+    eiv = [] # pandas object
+
+    iv_mean_omega_45 = [] # mean of all omega=45 for iv
+    eiv_mean_omega_45 = [] # mean of all omega=45 for eiv
+
+    iv_stdev_omega_45 = [] # stdev of all omega=45 for iv
+    eiv_stdev_omega_45 = [] # stdev of all omega=45 for eiv
+
+    sin_omega = []
+
+    table2 = [] # step before producing output table  iv table * sin(omega)
+    iv_ratio_omega_90 = [] # uses when calculating bank2 table2
+    mean_iv_ratio_omega_90 = [] # mean of iv ratio omega 90
+
+
+class VDriveHandler(object):
+
+    def __init__(self, filename=''):
+        self.data = Data()
+        self.bank1 = Bank()
+        self.bank2 = Bank()
+        self.mean_table = np.NaN
+        self.output_table = [] # final table of iv for output
+        self.filename = filename
+        self.__load_vdrive()
+
+    def run(self):
+        # full process
+        self.initialize_bank_xaxis()
+        self.keep_columns_of_interest()
+        self.isolating_banks()
+        self.calculate_mean_omega_45()
+        self.calculate_sin_omega()
+        self.calculate_bank2_iv_ratio_omega_90()
+        self.calculate_table2()
+        self.calculate_mean_table2()
+        self.calculate_output_table()
+
+    def calculate_sin_omega(self):
+        """sin(Pi()/180 * omega
+        this requires the initialize_bank_xaxis to be run before
+        """
+
+        #bank1
+        bank1_omega = self.bank1.omega
+        bank1_omega_rad = np.radians(bank1_omega)
+        bank1_sin_omega = np.sin(bank1_omega_rad)
+        self.bank1.sin_omega = bank1_sin_omega
+
+        # bank2
+        bank2_omega = self.bank2.omega
+        bank2_omega_rad = np.radians(bank2_omega)
+        bank2_sin_omega = np.sin(bank2_omega_rad)
+        self.bank2.sin_omega = bank2_sin_omega
+
+    def initialize_bank_xaxis(self):
+        self.initialize_bank1_xaxis()
+        self.initialize_bank2_xaxis()
+
+    def initialize_bank2_xaxis(self):
+        # initialization
+        omega = [np.int(_value) for _value in np.ones(12) * 45]
+
+        hrot1 = np.arange(0, 331, 30)
+        hrot2 = hrot1[::-1]
+
+        psi_init = np.ones(12) * 90
+        psi = [np.int(_value) for _value in psi_init]
+
+        phia = np.arange(180, -1, -30)
+        phib = np.arange(330, 180, -30)
+        phi1 = list(phia) + list(phib)
+        phi2 = phi1[::-1]
+
+        # full definition
+        full_omega = []
+        full_hrot = []
+        full_psi = []
+        full_phi = []
+
+        omega_offset = np.arange(0, 46, 5)
+        for _offset in omega_offset:
+            full_omega.append(omega + _offset)
+
+        hrot_offset = 0
+        while (hrot_offset < 5):
[…]
