### 2018-09-25 0d401e6
Added texture to list of branch to test
