### 2018-09-25 5fcc81b
Refs #23. update isntallation instruction
