### 2018-09-25 bd93049
Refs #23. updated installation instruction
