### 2018-09-25 7632704
Refs #23. minor wording changes
