### 2018-09-25 ornlneutronimaging/braggedgemodeling#23 @yxqd
@mikapfl @katyhuff I have updated the [installation instructions](https://ornlneutronimaging.github.io/braggedgemodeling/installation.html). It is now clarified that `conda` is the preferred installation method. The [using setup.py](https://ornlneutronimaging.github.io/braggedgemodeling/installation.html#using-setup-py) section is expanded with details, especially on installing [diffpy.structure python 3](https://github.com/diffpy/diffpy.structure/tree/python3) from source. Please check. Thanks.
