### 2018-09-25 63eb4b6
Refs openjournals/joss-reviews#973: tweaking letter case
