+++ JOSS-paper/paper.tex
@@ -4,7 +4,7 @@
 \usepackage{graphicx}
 
 
-\title{bem: Modeling for Neutron Bragg-Edge Imaging}
+\title{bem: modeling for neutron Bragg-edge imaging}
 \author[1]{Jiao Y. Y. Lin \thanks{Corresponding author. email: linjiao@ornl.gov}}
 \author[1]{Gian Song}
 \affil[1]{Neutron Scattering Division, Oak Ridge National Lab}
@@ -19,16 +19,16 @@
 Due to its zero net charge, neutron is a unique probe of materials.
 Low neutron absorption and scattering cross sections by most nuclei make it
 suitable for studying bulk samples.
-Unlike X-ray scattering, neutron form factors are not monotonically dependent of atomic numbers;
+Unlike X-ray scattering, neutron form factors are not monotonically dependent on atomic numbers;
 the fact that the neutron scattering cross section of hydrogen is large makes neutron
 a useful tool in biology.
-In the past half century, Neutron imaging has seen growing applications
+In the past half century, neutron imaging has seen growing applications
 in various scientific fields
 including physics, engineering sciences, biology, and archaeology
 \cite{strobl2009}.
 
 With energy-resolved neutron imaging techniques,
-Neutron Bragg-edge imaging has recently found applications for materials science in phase mapping,
+neutron Bragg-edge imaging has recently found applications for materials science in phase mapping,
 stress/strain mapping, and texture analysis
 \cite{lehmann2010, sato2017}.
 To model Bragg-edge neutron imaging data, it is necessary to calculate
@@ -43,7 +43,6 @@ the March-Dollase texture model, and the Jorgensen peak profile
 \cite{vogel2000thesis}.
 
 \section{Notice of Copyright}\label{notice-of-copyright}
-
 This manuscript has been authored by UT-Battelle, LLC under Contract No.
 DE-AC05-00OR22725 with the U.S. Department of Energy. The United States
 Government retains and the publisher, by accepting the article for
