### 2018-09-25 ornlneutronimaging/braggedgemodeling#23 @katyhuff
Weirdly, the setup.py instructions didn't work for me until I neglected --user in the diffpy.structure setup.py install. Probably something about my (admittedly, increasingly complex) various conda environments.

Works now, tests pass.
