### 2018-09-25 ornlneutronimaging/braggedgemodeling#23 @yxqd
Thanks @katyhuff for checking!
