### 2018-09-25 #973 @katyhuff
@yxqd This tool is small, certainly, but a functional utility. My only remaining concern is with the examples question.

Part of the review for JOSS is with regards to the understandability and usability of the code. Toward this aim, we ask the question: _Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)._

Indeed, as an exploring user, I immediately found and ran your tutorial.ipynb file. I even thought I understood what's going on it. However, it was pretty brief and I certainly wouldn't call it a tutorial or a "real world" example since the problems being approached aren't described at all in the notebook. Later, I saw that the corresponding page in the docs is the exact same example, but with just a few extra sentences to help the user: https://ornlneutronimaging.github.io/braggedgemodeling/tutorial.html . Even though this page is still very minimal, it is far easier to understand!

Accordingly, I'd like to suggest perhaps fleshing out, just a bit, these examples so that it's clearer who should use this and what analysis really calls for it. It would generally be ideal if the notebook in the source code included the same level of explanation as the readthedocs site.

Otherwise, this looks ok to me and I was able to run the tests and tutorial notebook successfully.
