### 2018-09-25 ornlneutronimaging/braggedgemodeling#24 @stuartcampbell
RuntimeWarnings when running the tutorial
When running through the tutorial, I saw the following RuntimeWarnings:
```
/anaconda3/envs/joss-review-bem/lib/python3.6/site-packages/bem/xtaloriprobmodel.py:57: RuntimeWarning: invalid value encountered in true_divide
  div = np.array(hkl, float)/np.array(hkl1, float)
/anaconda3/envs/joss-review-bem/lib/python3.6/site-packages/bem/xtaloriprobmodel.py:76: RuntimeWarning: invalid value encountered in arcsin
  alpha = np.pi/2 - np.arcsin(wavelen/2./peak.d)
/anaconda3/envs/joss-review-bem/lib/python3.6/site-packages/bem/xtaloriprobmodel.py:57: RuntimeWarning: divide by zero encountered in true_divide
  div = np.array(hkl, float)/np.array(hkl1, float)
/anaconda3/envs/joss-review-bem/lib/python3.6/site-packages/bem/xtaloriprobmodel.py:59: RuntimeWarning: invalid value encountered in multiply
  if np.allclose(hkl, np.array(hkl1)*div):
```

These were produced from calls to ` xscalculator.plotAll(wavelengths)`.
