### 2018-09-25 ornlneutronimaging/braggedgemodeling#25 @stuartcampbell
Add DOIs for JOSS Paper references
Add the DOIs for the references in the JOSS paper.
