### 2018-09-25 #973 @stuartcampbell
Hi @yxqd,
I installed using `conda`, worked without an issue.  Although as a small usability thing, I would change the install to be a single line: `$ conda install braggedgemodeling  -c condo-forge` - this means that you are not permanently. changing the default channels.

Running through the tutorials the answers look like something I would expect.  As a small suggestion, I would put images of the expected example plots in the tutorial so users can check that they are seeing what is expected.   I know the images are in the notebook, but as you point people at the website then I would put them there as well.  Also, maybe put a link to the notebook on the tutorial webpage to make it easier for people to find.

I did experience some RuntimeWarnings, I have created an issue for them (see https://github.com/ornlneutronimaging/braggedgemodeling/issues/24) but I don't think that they have to be fixed for this review.

I think overall the package looks good and the only outstanding issue I can see is that your references do not include the DOIs for them.  I've opened the issue https://github.com/ornlneutronimaging/braggedgemodeling/issues/25

After that is done I think all the requirements are satisfied.
