### 2018-09-26 #973 @katyhuff
Good catch @stuartcampbell ... I completely failed to notice the missing DOIs!
