### 2018-09-27 #973 @yxqd
Thank you @katyhuff @stuartcampbell for the comments. I will flesh out the notebook and the documentation as you suggested.
