### 2018-09-27 ornlneutronimaging/braggedgemodeling#26 @yxqd
improve documentation
See openjournals/joss-reviews#973

Improve notebook and sphinx tutorial
