### 2018-09-27 37dbc25
Refs #26. updated notebook with more instructions
