### 2018-09-27 5611de8
Refs #26. updated sphinx tutorial
