+++ docs/tutorial.rst
@@ -5,37 +5,62 @@ Tutorial
 
 Please see :ref:`installation` before start here.
 
+This tutorial also exists as a jupyter notebook: https://github.com/ornlneutronimaging/braggedgemodeling/blob/master/notebooks/tutorial.ipynb
+
+This tutorial walks through some basic steps of using braggedgemodeling for modeling of neutron Bragg edge spectrum.
+
+First we import some essential tools::
+
+  import os, numpy as np
+  import warnings
+  %matplotlib notebook
+  from matplotlib import pyplot as plt
+  from bem.matter import Atom, Lattice, Structure
+
 Atomic structure
 ----------------
 
-Let us start by creating an atomic structure::
+The first step is to create a model of the material. Here we use bcc Fe as the example. In the following cell, we create the atomic structure of bcc Fe by adding atoms to a lattice with a space group::
 
-  from bem.matter import Atom, Lattice, Structure
-  atoms = [Atom('Ni', (0,0,0)), Atom('Ni', (0.5, 0.5, 0)),
-           Atom('Ni', (0.5,0,0.5)), Atom('Ni', (0, 0.5, 0.5))]
-  a=3.5238
+  atoms = [Atom('Fe', (0,0,0)), Atom('Fe', (0.5, 0.5, 0.5))]
+  a=2.856
   alpha = 90.
   lattice = Lattice(a=a, b=a, c=a, alpha=alpha, beta=alpha, gamma=alpha)
-  fccNi = Structure(atoms, lattice, sgid=225)
+  astruct = Structure(atoms, lattice, sgid=229)
 
 .. note:: You can also use :code:`bem.matter.loadCif(path)` to load an atomic structure
-   from a CIF file.
+   from a CIF file::
 
-Cross section
--------------
+     from bem.matter import loadCif
+     astruct = loadCif(path)
 
-Then we can perform a simple Bragg Edge neutron cross section calculation and plot them::
+No texture
+----------
 
-  # define wavelength axis
-  import numpy as np
+It is always good to start with a simplified case. Here we consider a bcc Fe sample without texture, and the following code computes and plots various contributions to the neutron cross section for this material.
+
+The neutron Bragg edge spectrum is wavelength (𝜆) dependent, so we will calculate the cross sections as functions of  𝜆
+
+Define wavelength axis::
+
   wavelengths = np.arange(0.05, 5.5, 0.005)
-  T = 300
-  # create calculator
+
+Experimental condition::
+  
+  T = 300 # Kelvin
+  
+Calculate cross sections
+""""""""""""""""""""""""
+  
+Create calculator. Larger max_diffraction_index will lead to more accurate result at low wavelength::
+
   from bem import xscalc
-  xscalculator = xscalc.XSCalculator(fccNi, T, max_diffraction_index=4)
-  # compute various contributions
-  # In neutron Bragg Edge data analysis, it may not be necessary to calculate all these
-  # contributions, but it is useful to see them when exploring.
+  xscalculator = xscalc.XSCalculator(astruct, T, max_diffraction_index=4)
+
+Compute various contributions
+
+Note: in neutron Bragg Edge data analysis, it may not be necessary to calculate all these contributions, but it is useful to see them when exploring::
+
   coh_el_xs = xscalculator.xs_coh_el(wavelengths)
   inc_el_xs = xscalculator.xs_inc_el(wavelengths)
   abs_xs = xscalculator.xs_abs(wavelengths)
[…]
