### 2018-09-27 420d61e
Refs #26. added images for sphinx tutorial
