### 2018-09-27 e4a4842
Refs #26. minor tweaks of sphinx tutorial
