### 2018-09-27 0354ce5
Refs #26. minor tweaks of tutorial notebook
