### 2018-09-27 ba35c6e
Refs #26. added images to gh-pages
