### 2018-09-27 ed8c8b5
Refs #25. added DOIs for journal papers. cannot find DOI for Vogel thesis.
