+++ JOSS-paper/paper.bib
@@ -6,7 +6,8 @@
   number={24},
   pages={243001},
   year={2009},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0022-3727/42/24/243001}
 }
 
 @article{sato2017,
@@ -17,7 +18,8 @@
   number={1},
   pages={7},
   year={2017},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
+  doi={10.3390/jimaging4010007}
 }
 
 @article{lehmann2010,
@@ -28,12 +30,14 @@
   number={3},
   pages={515--522},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00339-010-5602-7}
 }
 
 @phdthesis{vogel2000thesis,
   title={A Rietveld-approach for the analysis of neutron time-of-flight transmission data},
   author={Vogel, Sven},
   year={2000},
-  school={Verlag nicht ermittelbar}
+  school={Verlag nicht ermittelbar},
+  doi={}
 }
