### 2018-09-27 #973 @yxqd
@katyhuff @stuartcampbell  documentation was updated as you suggested. Please check https://ornlneutronimaging.github.io/braggedgemodeling/tutorial.html and https://github.com/ornlneutronimaging/braggedgemodeling/blob/master/notebooks/tutorial.ipynb.

DOIs were added for journal papers referenced. I cannot find the DOI for the phd thesis referenced, however. Please check https://github.com/ornlneutronimaging/braggedgemodeling/blob/master/JOSS-paper/paper.bib

Thanks!
