### 2018-09-27 #973 @mikapfl
Hi,

unfortunately (for the review) I'm in vacations now and will only be able to continue with the review in two weeks.

Cheers,

Mika
--
Diese Nachricht wurde von meinem Android-Mobiltelefon mit K-9 Mail gesendet.
