### 2018-09-27 #973 @stuartcampbell
@yxqd the bib file looks good to me now.  @katyhuff does @yxqd need to either update the release version to make sure the correct .bib file and paper are in the released git tag ?
