### 2018-09-27 #973 @katyhuff
@stuartcampbell : An updated doi will be created in the final step before acceptance, if a new release is desired, to reflect review-related changes to the paper and code, many authors do issue a new (minor or patch) release version at that stage as well.

@yxqd  should wait until the acceptance to do that though, so that this task only needs to be done once. I would like to wait for @mikapfl's review before that stage. Though I understand it may be a couple more weeks from now, I hope that won't put an undue burden on your development process.
