### 2018-09-29 #973 @yxqd
@mikapfl thanks for the note.

@stuartcampbell thanks. I just closed ornlneutronimaging/braggedgemodeling#25.

@katyhuff thanks. sounds a good plan to me
