### 2018-10-04 3d062b6
Added missing scripts to run first part of code (before Matlab)
