### 2018-10-14 ornlneutronimaging/braggedgemodeling#23 @mikapfl
Hi,

works now for me, only left issue is that the last line, which currently reads
```
$ cd braggedgemodeling && python setup.py install --user
```
should probably read
```
$ cd braggedgemodeling && python3 setup.py install --user
```
to be consistent with the rest of the instructions, which also use python**3**

Cheers

Mika
