### 2018-10-14 ornlneutronimaging/braggedgemodeling#28 @mikapfl
xscalculator: Inconsistency of using float vs. array of floats
Hi,

I'm getting somewhat surprising NaNs from xscalculator if setting the wavelen to a float instead of an array of floats. Using this slightly adapted example from the tutorial:
```python
import os, numpy as np
import warnings
from bem.matter import Atom, Lattice, Structure
from bem import xscalc

atoms = [Atom('Fe', (0,0,0)), Atom('Fe', (0.5, 0.5, 0.5))]
a=2.856
alpha = 90.
lattice = Lattice(a=a, b=a, c=a, alpha=alpha, beta=alpha, gamma=alpha)
astruct = Structure(atoms, lattice, sgid=229)

from bem import xtaloriprobmodel as xopm
texture_model = xopm.MarchDollase()
texture_model.r[(0,1,1)] = 2
texture_model.beta[(0,1,1)] = np.deg2rad(60.)

xscalculator = xscalc.XSCalculator(astruct, 300, texture_model)

print('array: {}'.format(xscalculator.xs_coh_el(np.array([1.5]))))
print('float: {}'.format(xscalculator.xs_coh_el(1.5)))
```
The output is:
```
array: [20.56132678]
float: nan
```

I would have expected the float result to be the same, since the `xs_coh_el` documentation states that `wavelen` has the type `float, floats`.

Cheers

Mika
