### 2018-10-14 ornlneutronimaging/braggedgemodeling#29 @mikapfl
Consistently document units
Hi,

in many places (e.g. `xscalc.XSCalculator.xs_coh_el`) the input and output units are consistently documented; however, in other places, no unit documentation is provided, especially in `bem.matter` it seems that Ångström is the unit, but this is not documented. I think it would help understanding if all input and output units are documented for all objects and functions.

Cheers

Mika
