### 2018-10-14 ornlneutronimaging/braggedgemodeling#30 @mikapfl
Document bem.matter completely or provide pointers to diffpy documentation
Hi,

in the API documentation, `bem.matter` is not fully documented.
Specifically, `Atom` and `Lattice` are not documented outside of the tutorial at all, and `Structure` is documented with the single sentence `a wrapper for Structure method that injects “sg” data member`, which is not very helpful.

Please enhance the API documentation to include all parts of `bem.matter` which end-users are expected to use and with the information relevant to end users. Since most of `bem.matter` is only a very thin wrapper around diffpy, you could probably also include links to the relevant diffpy documentation instead of duplicating it.

Cheers

Mika
