### 2018-10-14 #973 @mikapfl
Hi,

I looked at the software and documentation today and found it generally easy to follow and the tutorial to be enough to get me started (I can't judge if this is close to a real-world analysis problem, because I don't know how complicated the real world is in the bragg-edge neutron diffraction camp). However, I found a few usability and documentation issued, which I have reported:
* ornlneutronimaging/braggedgemodeling/issues/28
* ornlneutronimaging/braggedgemodeling/issues/29
* ornlneutronimaging/braggedgemodeling/issues/30

None of them are show stoppers, but I think that new users could benefit more from the documentation and the API would be easier to use if they would be fixed.

I'll check the remaining items on the review to-do list tomorrow.

Cheers

Mika
