### 2018-10-15 #973 @mikapfl
Hi,

regarding the review question if all functional claims are implemented in the software, I'm having problems to understand if you actually implement
> Calculate diffraction peaks data (d-spacing, form factors, etc.) according to sample crystal structure

At least, this seems to be missing from the documentation (tutorial or API documentation) - how would I, say, calculate a form factor from a sample crystal structure? I haven't opened an issue in the bem issue tracker for this since I'm feeling like I'm missing something here, maybe you can comment, @yxqd ?

Cheers

Mika
