### 2018-10-15 #973 @mikapfl
Hi,

regarding the citations: The citations are linked appropriately and as far as I could see are also relevant. Only one thing for the citations to be done: I think the right citation for Sven Vogel's dissertation would be
`Vogel, S. (2000). A rietveld-approach for the analysis of neutron time-of-flight transmission data (PhD thesis). Retrieved from https://macau.uni-kiel.de/receive/dissertation_diss_00000330.`
At the moment it reads "Verlag nicht ermittelbar" which is just "publisher unknown" in German.

Cheers

Mika
