### 2018-10-15 2eddfd0
Refs #23. fixed typo
