### 2018-10-16 #973 @yxqd
Thanks, @mikapfl, for the comments. I will check on the issues.

Regarding "Calculate diffraction peaks data (d-spacing, form factors, etc.) according to sample crystal structure", this is done in the [bem.diffraction](https://github.com/ornlneutronimaging/braggedgemodeling/blob/master/bem/diffraction.py) module. d-spacing and form factor are essential to the Bragg-edge spectrum; without them the Bragg edge curves will not have any edges: each edge is located at a d-spacing position. The "jump" of intensities around the edge is mostly influenced by the form factor.

Thanks for finding the problem in citation, I will fix it.
