### 2018-10-16 827e22d
Refs #28. fixing the bug and adding a test
