### 2018-10-16 ornlneutronimaging/braggedgemodeling#29 @yxqd
Thanks, @mikapfl. More description on units were added to https://ornlneutronimaging.github.io/braggedgemodeling/api.html. Regarding `bem.matter`, I will comment in #30
