### 2018-10-16 ornlneutronimaging/braggedgemodeling#30 @yxqd
The `bem.matter.loadCif` is really the API intended to be used in reality, since in most cases a cif file is available from the user, or from a crystallography database. I will make this clear and add refs to diffpy doc.
