### 2018-10-16 41919a5
Refs #30. improving docs for dem.matter.Structure
