### 2018-10-16 87917d3
Refs #30. fine-tuning
