### 2018-10-16 37d5efb
Refs #30. fine-tuning
