### 2018-10-16 ornlneutronimaging/braggedgemodeling#28 @yxqd
@mikapfl thanks for catching the bug. Could you please check the fix? Thanks.
