### 2018-10-16 ornlneutronimaging/braggedgemodeling#30 @yxqd
@mikapfl please check [the new API doc](https://ornlneutronimaging.github.io/braggedgemodeling/api.html#atomic-structure).
