### 2018-10-16 ea36b27
Refs #29, #30. improve docs
