### 2018-10-16 b6cc456
Refs openjournals/joss-reviews#973. fixing ref to vogel thesis
