+++ JOSS-paper/paper.bib
@@ -38,6 +38,7 @@
   title={A Rietveld-approach for the analysis of neutron time-of-flight transmission data},
   author={Vogel, Sven},
   year={2000},
-  school={Verlag nicht ermittelbar},
-  doi={}
+  school={The University of Kiel},
+  url={https://core.ac.uk/download/pdf/56723242.pdf},
+  urldate={2018-10-15}
 }
