### 2018-10-16 #973 @yxqd
@mikapfl Please review ornlneutronimaging/braggedgemodeling#23, ornlneutronimaging/braggedgemodeling#28, ornlneutronimaging/braggedgemodeling#29, ornlneutronimaging/braggedgemodeling#30. Thanks.
