### 2018-10-16 ornlneutronimaging/braggedgemodeling#23 @mikapfl
Looks good for me now!
