### 2018-10-16 ornlneutronimaging/braggedgemodeling#30 @mikapfl
Looks good, thanks!
