### 2018-10-16 #973 @mikapfl
@yxqd `bem.diffraction` seems to be entirely undocumented. I understand that you need it mainly "behind the scenes" to calculate the full Bragg edge curve, so maybe you should not put "Calculate diffraction peaks data (d-spacing, form factors, etc.) according to sample crystal structure" as the first (and thus implied to be most important) feature in the description of features in the README. If users of this package are mainly interested in Bragg edge curves, put those front and center in your description of features. You can, if you want to, still mention that you of course also calculate diffraction peak data to be able to calculate a Bragg edge curve, but I think you should somehow make clear that it is rather an internal thing.

Cheers

Mika
