### 2018-10-16 #973 @katyhuff
I believe the concern regarding references has been handled. This leaves just the final check boxes by @mikapfl regarding functionality and documentation of the API. Once those are handled, we should be good to go! @yxqd I hope you'll let us know when you believe these have been handled.
