### 2018-10-16 8b04ae0
Refs openjournals/joss-reviews#973. revising "features"
