### 2018-10-16 #973 @yxqd
@mikapfl @katyhuff  I reworded the "features" to hide the details about diffraction calculation. Please check.
