### 2018-10-17 #973 @mikapfl
Hi,

thanks, that solves the remaining issue for me, from my side, everything has been handled now. (-:

Cheers

Mika
