### 2018-10-18 #973 @katyhuff
Excellent! Thank you @mikapfl and @stuartcampbell for the reviews! I know it's a lot of work, and I think you both conducted very helpful and thorough.

Thank you @yxqd for this interesting submission, and for engaging actively in the review process.
Could you please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission! (You may want to be sure to double check the paper, all minor details, etc. before creating the archive).
