### 2018-10-19 #973 @mikapfl
Hi,

the "cite as" part in http://joss.theoj.org/papers/10.21105/joss.00973 looks weird, is that intended?

Cheers

Mika
