### 2018-10-19 #973 @arfon
Nope. That's a bug. I'll fix early next week.

On Fri, Oct 19, 2018 at 07:58, Mika Pflüger <notifications@github.com>
wrote:

> Hi,
>
> the "cite as" part in http://joss.theoj.org/papers/10.21105/joss.00973
> looks weird, is that intended?
>
> Cheers
>
> Mika
>
